# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClashLayoutsHub API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /comments/{baseId}:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments for a base
      parameters:
        - name: baseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Comments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"

    post:
      operationId: addComment
      tags: [comments]
      summary: Add a comment to a base
      parameters:
        - name: baseId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentBody"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      summary: List all reports (admin only)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Reports list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportsListResponse"

    post:
      operationId: createReport
      tags: [reports]
      summary: Report a base issue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReportBody"
      responses:
        "201":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List submissions (admin only)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Submissions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionsListResponse"

    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit a base for review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmissionBody"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /submissions/{id}:
    put:
      operationId: updateSubmission
      tags: [submissions]
      summary: Approve/reject a submission (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubmissionBody"
      responses:
        "200":
          description: Submission updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Comment:
      type: object
      properties:
        id:
          type: string
        base_id:
          type: string
        user_name:
          type: string
        content:
          type: string
        rating:
          type: integer
        approved:
          type: boolean
        created_at:
          type: string
      required:
        - id
        - base_id
        - user_name
        - content
    CreateCommentBody:
      type: object
      properties:
        user_name:
          type: string
        content:
          type: string
        rating:
          type: integer
      required:
        - user_name
        - content
    Report:
      type: object
      properties:
        id:
          type: string
        base_id:
          type: string
        base_title:
          type: string
        reason:
          type: string
        reporter_ip:
          type: string
        created_at:
          type: string
      required:
        - id
        - base_id
        - reason
    CreateReportBody:
      type: object
      properties:
        base_id:
          type: string
        reason:
          type: string
      required:
        - base_id
        - reason
    ReportsListResponse:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: "#/components/schemas/Report"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - reports
        - total
        - page
        - totalPages
    Submission:
      type: object
      properties:
        id:
          type: string
        townhall:
          type: integer
        base_type:
          type: string
        layout_link:
          type: string
        description:
          type: string
        image_url:
          type: string
        status:
          type: string
        admin_notes:
          type: string
        submitted_at:
          type: string
      required:
        - id
        - layout_link
        - status
    CreateSubmissionBody:
      type: object
      properties:
        townhall:
          type: integer
        base_type:
          type: string
        layout_link:
          type: string
        description:
          type: string
        image_url:
          type: string
      required:
        - layout_link
    UpdateSubmissionBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected]
        admin_notes:
          type: string
      required:
        - status
    SubmissionsListResponse:
      type: object
      properties:
        submissions:
          type: array
          items:
            $ref: "#/components/schemas/Submission"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - submissions
        - total
        - page
        - totalPages
...