# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClashLayoutsHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: string;
  base_id: string;
  user_name: string;
  content: string;
  rating?: number;
  approved?: boolean;
  created_at?: string;
}