# lib/api-zod/src/generated/types/createSubmissionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClashLayoutsHub API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSubmissionBody {
  townhall?: number;
  base_type?: string;
  layout_link: string;
  description?: string;
  image_url?: string;
}