# lib/api-zod/src/generated/types/submissionsListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClashLayoutsHub API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Submission } from "./submission";

export interface SubmissionsListResponse {
  submissions: Submission[];
  total: number;
  page: number;
  totalPages: number;
}